Refactor VisualRecognition services into a separate namespace/folder
