+++ src/VisualRecognition/Mappers/ClassifierMapper.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using VisualRecognition.ViewModels;
-using WatsonServices.Models;
+using WatsonServices.Models.VisualRecognition;
 
 namespace VisualRecognition.Mappers
 {
