+++ src/VisualRecognition/Mappers/ScoresMapper.cs
@@ -1,5 +1,5 @@
 ﻿using VisualRecognition.ViewModels;
-using WatsonServices.Models;
+using WatsonServices.Models.VisualRecognition;
 
 namespace VisualRecognition.Mappers
 {
