+++ src/WatsonServices/Models/VisualRecognition/Classifier.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace WatsonServices.Models
+namespace WatsonServices.Models.VisualRecognition
 {
     public class Classifier
     {
