+++ src/WatsonServices/Models/VisualRecognition/ClassifiersResponse.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace WatsonServices.Models
+namespace WatsonServices.Models.VisualRecognition
 {
     public class ClassifiersResponse
     {
